# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance tracking API for managing accounts, transactions, budgets, and more.",
        "title": "Fortuna API",
        "contact": {
            "name": "API Support",
            "email": "support@fortuna.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api/v1",
    "paths": {
...
        "/recurring-templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all recurring templates for the workspace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring Templates"
                ],
                "summary": "List all recurring templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TemplateListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new recurring template with projection generation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring Templates"
                ],
                "summary": "Create a recurring template",
                "parameters": [
                    {
                        "description": "Template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/recurring-templates/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a single recurring template by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring Templates"
                ],
                "summary": "Get a recurring template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TemplateResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a recurring template and recalculates projections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recurring Templates"
                ],
                "summary": "Update a recurring template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a recurring template and all its projections",
                "tags": [
                    "Recurring Templates"
                ],
                "summary": "Delete a recurring template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateTemplateRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "amount": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "linkTransactionId": {
                    "type": "integer"
                },
                "settlementIntent": {
                    "description": "For CC accounts: \"immediate\" or \"deferred\"",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
...
        "handler.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ValidationError"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "handler.TemplateListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TemplateResponse"
                    }
                }
            }
        },
        "handler.TemplateResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "amount": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "settlementIntent": {
                    "description": "For CC accounts: \"immediate\" or \"deferred\"",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workspaceId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateTemplateRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "integer"
                },
                "amount": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "settlementIntent": {
                    "description": "For CC accounts: \"immediate\" or \"deferred\"",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "handler.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "For JWT: \"Bearer \u003cjwt_token\u003e\". For API tokens: just enter \"fort_xxx\" (Bearer prefix optional).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}